Fix IOCTL_BTHX_READ/WRITE_HCI parameter sizes
WRITE_HCI's second parameter is a ULONG, not a BTHX_HCI_PACKET_TYPE. These two types may or may not be of equal size, depending on the compiler.

READ_HCI is specified as taking a UCHAR representing the type. This driver
instead uses an enum type, with an implementation-defined size. Properly convert to this type from the specified UCHAR, and don't rely on the caller passing an enum-sized buffer.