Update WPP trace GUID for pcidrv and serial samples
It is common for people to copy and rename WDK sample when starting a new driver project. During the process sometimes people forgets to change the new driver's WPP trace GUID, causing difficulty when collecting trace logs. This happens for some inbox drivers as well. Note this is not a problem when using WDK project wizard in VS which always creates a new trace GUID.

The fix updates the trace GUID for pcidrv and serial samples, so that they won't collide with the inbox driver.