Update README.md
I added statements about how to change Platform Toolset of the projects in Step 2-i.
When I used VS2017 and VS2019 to try to build the projects, Driver Files wasn't shown in the menu. 
So I think it is necessary to add these steps for people who download these good sample projects.