Improvements to WFP Inspect Sample (#516)
Changes a `NT_ASSERT` to match what the comment above is saying, the previous version caused a system crash if the packet was not inbound and had a auth decision made already. 
The comment
`         // Skip pended connections in the list, for which the auth decision is already taken.`
`         // They should not be for inbound connections.`
to my understanding means inbound connections should be ignored and ignore connections for which the auth decision was already made. The latter is checked by:
` if (packet->authConnectDecision == 0)` 
which keeps `found` equal to `FALSE` triggering the packet and list entry to be set to `NULL` here:
`if (!found)
         {
            listEntry = NULL;
            packet = NULL;
         }`

Change also updates the .sln and .vxproj  files so the sample compiles on VS2019 without any changes - addressing #366 

Co-authored-by: Elaine Houha <elhouha@microsoft.com>