Removing Unused variable from SimpleAudioSample (#710)
There was a variable (MicArray_180_Degrees) being initialized but not used which was causing a warning and a subsequent failure when building the driver in the Windows Driver Kit TP process. This was fixed by removing the variable's initialization. Validated by building the driver and making sure it still installs and works correctly.