[input/kbfiltr] Remove Boilerplate Text From Callout in Sample Keyboard Input Driver Doc (#1165)
The current version of the `README.md` file for the keyboard filter
driver sample erroneously includes what looks like boilerplate text
in the callout at the bottom of the file. This text explains to the
writer of the documentation what to include in the callout text, and
should therefore have been removed when the actual content was added.

This change simply removes this change from the callout text.