Update ntdisp.c
Replace IoCreateDevice to IoCreateDeviceSecure. Documentation of IoCreateDevice mentions as follows. This sample is creating a named device object, and it is not specified security descriptor in the INF file. Thus, it should use IoCreateDeviceSecure.
---------------------------
IoCreateDevice can only be used to create an unnamed device object, or a named device object for which a security descriptor is set by an INF file. Otherwise, drivers must use IoCreateDeviceSecure to create named device objects. 
---------------------------